WordPress API/Bulk Fetch Articles By Slug/requirements.txt
httpx>=0.23.0,<0.24.0
loguru>=0.6.0,<0.7.0
python-dotenv>=0.20.0,<0.21.0
pytz==2022.1
yaspin>=2.2.0,<3.0.0
semantic_version
nanoid>=2.0.0,<3.0.0
jsonlines>=3.1.0,<4.0.0
pytest>=7.2.2,<7.3.0
